queue: harden concurrent queue lifecycle
Why:
Ensure cancellation, producer teardown, and quiescence remain correct
under adversarial thread interleavings and rejected configurations.

Impact:
Prevents stranded credits, stale rollback, reordering, and error leaks.

Before/After:
Before, several writer and producer races could strand queue ownership.
After, operation references and control transitions are helpable and exact.

Technical Overview:
Pack producer close state and transient references into one atomic word.
Expose BBQ reservations before advancing the visible publication tail.
Allow producers and consumers to promote reserved cells safely.
Bound each claim scan and preserve stable-producer publication order.
Serialize complete quiesce and destroy calls through a control owner.
Make reservation, lease, and builder cleanup non-failing in both cores.
Restore sparse queue flags after every timed-out control transition.
Release sparse disposal callbacks outside lane-local storage gates.
Close queue-configuration and constructed-main error-path leaks.
Add deterministic weak-memory, helping, timeout, and cleanup tests.

With the help of AI-Agents: Codex

+++ runtime/concurrent_array.c
@@ -151,19 +151,19 @@ void ca_capacity_read(const ca_queue_t *queue, ca_capacity_snapshot_t *snapshot)
     queue->ops->capacity_read(queue, snapshot);
 }
 
-uint32_t ca_epoch(const ca_queue_t *queue) {
+uint64_t ca_epoch(const ca_queue_t *queue) {
     return queue == NULL ? 0 : queue->ops->epoch(queue);
 }
 
-ca_status_t ca_wait_epoch(ca_queue_t *queue, uint32_t observed, const struct timespec *deadline) {
+ca_status_t ca_wait_epoch(ca_queue_t *queue, uint64_t observed, const struct timespec *deadline) {
     return queue == NULL ? CA_INVALID : queue->ops->wait_epoch(queue, observed, deadline);
 }
 
-uint32_t ca_capacity_epoch(const ca_queue_t *queue) {
+uint64_t ca_capacity_epoch(const ca_queue_t *queue) {
     return queue == NULL ? 0 : queue->ops->capacity_epoch(queue);
 }
 
-ca_status_t ca_wait_capacity_epoch(ca_queue_t *queue, uint32_t observed, const struct timespec *deadline) {
+ca_status_t ca_wait_capacity_epoch(ca_queue_t *queue, uint64_t observed, const struct timespec *deadline) {
     return queue == NULL ? CA_INVALID : queue->ops->wait_capacity_epoch(queue, observed, deadline);
 }
